phase prise temps crud done

+++ database/migrations/2024_05_09_131648_create_phase_gachages_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('phase_gachages', function (Blueprint $table) {
             $table->id();
-            $table->integer('temperature')->nullable();
+            $table->float('temperature')->nullable();
             $table->float('temperature_salle')->nullable();
             $table->float('humidite')->nullable();
             $table->float('p_prisme')->nullable();
